Anonymous iframe: Stop using unpartitionned cookies.
Stop feeding anonymous iframe document with unpartitionned cookies.

When the CookiePartitionKey contains a nonce, it is from an
anonymous/fenced iframe. As a result, we must only take cookies from
this CookiePartitionKey and not interact with the unpartitionned ones.

A future patch will force every cookies defined from within the
anonymous iframe to be partitionned. See:
https://bugs.chromium.org/p/chromium/issues/detail?id=1275465

Doc:
https://docs.google.com/document/d/1ymsMVb6jo8vwwZj5gQAHakdoY2z_XxtCI6-FkmPFSK8/edit

Bug:1275464
Change-Id: I4b8499bd7b77120363848eb5389eff7d2a8d090e