[@container] Don't reattach fieldset while being laid out
Fieldsets were fully re-attached when any child were re-attached to
change the box structure for rendered legends. If the fieldset is a
container for size container queries, this meant the LayoutObject for
the fieldset would be destroyed while being laid out.

Instead, re-attach all descendants of fieldset, not the fieldset itself.

Bug: 1275087
Change-Id: I4bee17065828e97f46f920eb1077581b33e78d7e